+++ tket/test/src/test_GreedyPauli.cpp
@@ -20,7 +20,7 @@
 #include "tket/Circuit/PauliExpBoxes.hpp"
 #include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Gate/SymTable.hpp"
-#include "tket/PauliGraph/PauliGraph.hpp"
+#include "tket/Ops/ClassicalOps.hpp"
 #include "tket/Predicates/PassGenerators.hpp"
 #include "tket/Transformations/GreedyPauliOptimisation.hpp"
 #include "tket/Utils/Expression.hpp"
@@ -28,36 +28,20 @@
 namespace tket {
 namespace test_GreedyPauliSimp {
 
-SCENARIO("Unsupported circuits") {
-  GIVEN("Circuit with mid-circ measurements") {
-    Circuit circ(2, 2);
-    circ.add_op<unsigned>(OpType::H, {0});
-    circ.add_op<unsigned>(OpType::Rx, 0.5, {1});
-    circ.add_op<unsigned>(OpType::Measure, {0, 0});
-    circ.add_op<unsigned>(OpType::CX, {0, 1});
-    REQUIRE_THROWS_MATCHES(
-        Transforms::greedy_pauli_optimisation().apply(circ),
-        MidCircuitMeasurementNotAllowed,
-        MessageContains(
-            "PauliGraph does not support mid-circuit measurements"));
-  }
-  GIVEN("Circuit with resets") {
+SCENARIO("Exception handling") {
+  GIVEN("Invalid arguments") {
     Circuit circ(1);
-    circ.add_op<unsigned>(OpType::H, {0});
-    circ.add_op<unsigned>(OpType::Reset, {0});
     REQUIRE_THROWS_MATCHES(
-        Transforms::greedy_pauli_optimisation().apply(circ), BadOpType,
-        MessageContains("Cannot add gate to PauliGraph"));
-  }
-  GIVEN("Circuit with conditional gates") {
-    Circuit circ(2, 2);
-    circ.add_conditional_gate<unsigned>(OpType::Rz, {0.5}, {0}, {0}, 0);
+        Transforms::greedy_pauli_optimisation(0.3, 0.3, 0, 10).apply(circ),
+        Transforms::GreedyPauliSimp::GreedyPauliSimpError,
+        MessageContains("max_lookahead must be greater than 0."));
     REQUIRE_THROWS_MATCHES(
-        Transforms::greedy_pauli_optimisation().apply(circ), BadOpType,
-        MessageContains(
-            "Can only make a PauliGraph from a circuit of basic gates"));
+        Transforms::greedy_pauli_optimisation(0.3, 0.3, 10, 0).apply(circ),
+        Transforms::GreedyPauliSimp::GreedyPauliSimpError,
+        MessageContains("max_tqe_candidates must be greater than 0."));
   }
 }
+
 SCENARIO("Clifford synthesis") {
   GIVEN("Empty circuit") {
     Circuit circ(3);
@@ -129,6 +113,26 @@ SCENARIO("Clifford synthesis") {
     Circuit d(circ);
     REQUIRE(Transforms::greedy_pauli_optimisation().apply(d));
     REQUIRE(test_unitary_comparison(circ, d, true));
+  }
+  GIVEN("Test search limits") {
+    Circuit circ(4);
+    circ.add_op<unsigned>(OpType::X, {0});
+    circ.add_op<unsigned>(OpType::SWAP, {1, 2});
+    circ.add_op<unsigned>(OpType::CX, {0, 2});
+    circ.add_op<unsigned>(OpType::H, {3});
+    circ.add_op<unsigned>(OpType::CZ, {1, 3});
+    circ.add_op<unsigned>(OpType::H, {2});
+    circ.add_op<unsigned>(OpType::CX, {3, 2});
+    circ.add_op<unsigned>(OpType::Z, {2});
+    circ.add_op<unsigned>(OpType::SWAP, {3, 1});
+    circ.add_op<unsigned>(OpType::CY, {0, 2});
+    Circuit d1(circ);
+    Circuit d2(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 2, 1).apply(d1));
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 20, 20).apply(d2));
+    REQUIRE(test_unitary_comparison(circ, d1, true));
+    REQUIRE(test_unitary_comparison(circ, d2, true));
+    REQUIRE(d1 != d2);
   }
 }
 SCENARIO("Complete synthesis") {
@@ -213,6 +217,55 @@ SCENARIO("Complete synthesis") {
     REQUIRE(Transforms::greedy_pauli_optimisation().apply(d));
     REQUIRE(test_unitary_comparison(circ, d, true));
   }
+  GIVEN("5Q PauliExp Circuit with search limits") {
+    Circuit circ(5);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::X, Pauli::X}, 0.3)),
+        {0, 1, 4});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::Z, Pauli::Y}, -0.1)),
+        {2, 3, 0});
+    circ.add_box(
+        PauliExpPairBox(
+            SymPauliTensor({Pauli::X, Pauli::Z, Pauli::Y}, 1.0),
+            SymPauliTensor({Pauli::Z, Pauli::X, Pauli::Y}, 0.4)),
+        {0, 2, 4});
+    circ.add_box(
+        PauliExpCommutingSetBox({
+            {{
+                 Pauli::I,
+                 Pauli::Y,
+                 Pauli::I,
+             },
+             -0.1},
+            {{Pauli::X, Pauli::Y, Pauli::Z}, -1.2},
+            {{Pauli::X, Pauli::Y, Pauli::Z}, 0.5},
+        }),
+        {1, 2, 3});
+    circ.add_box(
+        PauliExpCommutingSetBox({
+            {{
+                 Pauli::I,
+                 Pauli::X,
+                 Pauli::I,
+             },
+             -0.15},
+            {{Pauli::X, Pauli::X, Pauli::Z}, -1.25},
+            {{Pauli::X, Pauli::X, Pauli::Z}, 0.2},
+        }),
+        {0, 3, 4});
+    circ.add_op<unsigned>(OpType::CX, {0, 2});
+    circ.add_op<unsigned>(OpType::SWAP, {2, 3});
+    circ.add_op<unsigned>(OpType::H, {3});
+    circ.add_op<unsigned>(OpType::CZ, {1, 3});
+    Circuit d1(circ);
+    Circuit d2(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 3, 3).apply(d1));
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 30, 30).apply(d2));
+    REQUIRE(test_unitary_comparison(circ, d1, true));
+    REQUIRE(test_unitary_comparison(circ, d2, true));
+    REQUIRE(d1 != d2);
+  }
   GIVEN("Circuit with trivial Pauli exps") {
     Circuit circ(4);
     circ.add_box(PauliExpBox(SymPauliTensor({Pauli::X, Pauli::X}, 2)), {0, 1});
@@ -275,6 +328,385 @@ SCENARIO("Complete synthesis") {
     }
     REQUIRE(Transforms::greedy_pauli_optimisation().apply(g));
     REQUIRE(d == g);
+  }
+  GIVEN("Circuit with conditional gates") {
+    Circuit circ(2, 2);
+    circ.add_op<unsigned>(OpType::CX, {0, 1});
+    circ.add_conditional_gate<unsigned>(OpType::Rz, {0.5}, {0}, {0}, 0);
+    circ.add_op<unsigned>(OpType::CX, {0, 1});
+    Circuit d(2, 2);
+    Op_ptr cond = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::I}, 0.5)),
+        1, 0);
+    d.add_conditional_gate<unsigned>(OpType::Sdg, {}, {0}, {0}, 0);
+    d.add_conditional_gate<unsigned>(OpType::Z, {}, {0}, {0}, 0);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with conditional gates 2") {
+    Circuit circ(2, 1);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::Z}, 0.12)), {0, 1});
+    Op_ptr cond = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::Z}, 0.5)),
+        1, 0);
+    circ.add_op<unsigned>(cond, {0, 0, 1});
+    // two boxes anti-commute hence simultaneous diagonalisation
+    Circuit d(2, 1);
+    d.add_op<unsigned>(OpType::CY, {1, 0});
+    d.add_op<unsigned>(OpType::Rx, 0.12, {0});
+    d.add_conditional_gate<unsigned>(OpType::Sdg, {}, {0}, {0}, 0);
+    d.add_conditional_gate<unsigned>(OpType::Z, {}, {0}, {0}, 0);
+    d.add_op<unsigned>(OpType::CY, {1, 0});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with conditional gates and measures") {
+    Circuit circ(2, 2);
+    Op_ptr cond1 = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::X}, 0.5)),
+        1, 0);
+    Op_ptr cond2 = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::Y}, 0.12)),
+        1, 0);
+    circ.add_op<unsigned>(cond1, {0, 0, 1});
+    circ.add_op<unsigned>(OpType::Measure, {0, 0});
+    // can commute to the front
+    circ.add_op<unsigned>(cond2, {1, 0, 1});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ.count_n_qubit_gates(2) == 3);
+  }
+
+  GIVEN("Conditionals merging") {
+    Circuit circ(2, 2);
+    Op_ptr cond1 = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::X}, 0.25)),
+        1, 0);
+    Op_ptr cond2 = std::make_shared<Conditional>(
+        std::make_shared<PauliExpBox>(
+            SymPauliTensor({Pauli::Z, Pauli::X}, -0.25)),
+        1, 0);
+    circ.add_op<unsigned>(cond1, {0, 0, 1});
+    circ.add_op<unsigned>(OpType::Rz, 0.3, {0});
+    circ.add_op<unsigned>(cond2, {0, 0, 1});
+    circ.add_conditional_gate<unsigned>(OpType::CX, {}, {0, 1}, {0}, 0);
+    circ.add_conditional_gate<unsigned>(OpType::CX, {}, {0, 1}, {0}, 0);
+    circ.add_op<unsigned>(OpType::Rz, -0.3, {0});
+    // should all be canceled
+    Circuit d(2, 2);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with classical gates") {
+    Circuit circ(1, 4);
+    circ.add_op<unsigned>(OpType::H, {0});
+    circ.add_op<unsigned>(ClassicalX(), {1});
+    circ.add_op<unsigned>(ClassicalCX(), {0, 1});
+    circ.add_op<unsigned>(AndWithOp(), {2, 3});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with WASMs") {
+    std::string wasm_file = "string/with/path/to/wasm/file";
+    std::string wasm_func = "stringNameOfWASMFunc";
+    std::vector<unsigned> uv = {2, 1};
+    const std::shared_ptr<WASMOp> wop_ptr =
+        std::make_shared<WASMOp>(6, 1, uv, uv, wasm_func, wasm_file);
+    Circuit circ(1, 7);
+    circ.add_op<unsigned>(OpType::X, {0});
+    circ.add_op<UnitID>(
+        wop_ptr,
+        {Bit(0), Bit(1), Bit(2), Bit(3), Bit(4), Bit(5), WasmState(0)});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with mid-circuit measurements") {
+    Circuit circ(2, 2);
+    circ.add_op<unsigned>(OpType::T, {0});
+    circ.add_op<unsigned>(OpType::Measure, {0, 0});
+    circ.add_op<unsigned>(OpType::Tdg, {0});
+    circ.add_op<unsigned>(OpType::Measure, {1, 1});
+    Circuit d(2, 2);
+    d.add_op<unsigned>(OpType::Measure, {0, 0});
+    d.add_op<unsigned>(OpType::Measure, {1, 1});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+  GIVEN("Circuit with mid-circuit measurements 2") {
+    // -X
+    Circuit c1(1, 1);
+    c1.add_op<unsigned>(OpType::Z, {0});
+    c1.add_op<unsigned>(OpType::H, {0});
+    c1.add_op<unsigned>(OpType::Measure, {0, 0});
+    c1.add_op<unsigned>(OpType::T, {0});
+    Circuit d1(1, 1);
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    d1.add_op<unsigned>(OpType::Measure, {0, 0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::Rx, 3.75, {0});
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c1));
+    REQUIRE(c1 == d1);
+    // Y
+    Circuit c2(1, 1);
+    c2.add_op<unsigned>(OpType::V, {0});
+    c2.add_op<unsigned>(OpType::Measure, {0, 0});
+    c2.add_op<unsigned>(OpType::T, {0});
+    Circuit d2(1, 1);
+    d2.add_op<unsigned>(OpType::V, {0});
+    d2.add_op<unsigned>(OpType::Measure, {0, 0});
+    d2.add_op<unsigned>(OpType::Vdg, {0});
+    d2.add_op<unsigned>(OpType::Ry, 0.25, {0});
+    d2.add_op<unsigned>(OpType::V, {0});
+    // Vdg;Ry(0.25);V = T
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c2));
+    REQUIRE(c2 == d2);
+    // -Y
+    Circuit c3(1, 1);
+    c3.add_op<unsigned>(OpType::Vdg, {0});
+    c3.add_op<unsigned>(OpType::Measure, {0, 0});
+    c3.add_op<unsigned>(OpType::T, {0});
+    Circuit d3(1, 1);
+    d3.add_op<unsigned>(OpType::Vdg, {0});
+    d3.add_op<unsigned>(OpType::Measure, {0, 0});
+    d3.add_op<unsigned>(OpType::V, {0});
+    d3.add_op<unsigned>(OpType::Ry, 3.75, {0});
+    d3.add_op<unsigned>(OpType::V, {0});
+    d3.add_op<unsigned>(OpType::X, {0});
+    // V;Ry(3.75);V;X = T
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c3));
+    REQUIRE(c3 == d3);
+    // -Z
+    Circuit c4(1, 1);
+    c4.add_op<unsigned>(OpType::X, {0});
+    c4.add_op<unsigned>(OpType::Measure, {0, 0});
+    c4.add_op<unsigned>(OpType::T, {0});
+    Circuit d4(1, 1);
+    d4.add_op<unsigned>(OpType::X, {0});
+    d4.add_op<unsigned>(OpType::Measure, {0, 0});
+    d4.add_op<unsigned>(OpType::X, {0});
+    d4.add_op<unsigned>(OpType::Rz, 3.75, {0});
+    d4.add_op<unsigned>(OpType::X, {0});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c4));
+    REQUIRE(c4 == d4);
+  }
+
+  GIVEN("Circuit with resets") {
+    Circuit circ(2);
+    circ.add_op<unsigned>(OpType::CX, {0, 1});
+    circ.add_op<unsigned>(OpType::Reset, {0});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ == d);
+  }
+
+  GIVEN("Circuit with resets 2") {
+    // -X/Z
+    Circuit c1(1);
+    c1.add_op<unsigned>(OpType::Z, {0});
+    c1.add_op<unsigned>(OpType::H, {0});
+    c1.add_op<unsigned>(OpType::Reset, {0});
+    Circuit d1(1);
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    d1.add_op<unsigned>(OpType::Reset, {0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::H, {0});
+    d1.add_op<unsigned>(OpType::X, {0});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c1));
+    REQUIRE(c1 == d1);
+
+    // X/-Z
+    Circuit c2(1);
+    c2.add_op<unsigned>(OpType::X, {0});
+    c2.add_op<unsigned>(OpType::H, {0});
+    c2.add_op<unsigned>(OpType::Reset, {0});
+    Circuit d2(1);
+    d2.add_op<unsigned>(OpType::H, {0});
+    d2.add_op<unsigned>(OpType::Z, {0});
+    d2.add_op<unsigned>(OpType::Reset, {0});
+    d2.add_op<unsigned>(OpType::Z, {0});
+    d2.add_op<unsigned>(OpType::H, {0});
+    d2.add_op<unsigned>(OpType::H, {0});
+    d2.add_op<unsigned>(OpType::Z, {0});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(c2));
+    REQUIRE(c2 == d2);
+  }
+
+  GIVEN("Circuit with measures, classicals, and resets") {
+    Circuit circ(3, 1);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::Z, Pauli::Z}, 0.3)),
+        {0, 1, 2});
+    circ.add_op<unsigned>(OpType::Measure, {0, 0});
+    circ.add_op<unsigned>(ClassicalX(), {0});
+    circ.add_op<unsigned>(OpType::Reset, {1});
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(circ.count_n_qubit_gates(2) == 4);
+    REQUIRE(circ.count_gates(OpType::ClassicalTransform) == 1);
+    REQUIRE(circ.count_gates(OpType::Measure) == 1);
+    REQUIRE(circ.count_gates(OpType::Reset) == 1);
+  }
+  GIVEN("Compile to ZZPhase") {
+    Circuit circ(2);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::X}, 0.3)), {0, 1});
+    Circuit d1(circ);
+    Circuit d2(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 500, 500, 0, true)
+                .apply(d1));
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 500, 500, 0, false)
+                .apply(d2));
+    REQUIRE(test_unitary_comparison(circ, d1, true));
+    REQUIRE(test_unitary_comparison(circ, d2, true));
+    REQUIRE(d1.count_n_qubit_gates(2) == 1);
+    REQUIRE(d2.count_n_qubit_gates(2) == 2);
+  }
+  GIVEN("Multiple ZZPhases at once") {
+    Circuit circ(6);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::X}, 0.3)), {0, 1});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::X}, 0.1)), {2, 3});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::Y}, 0.2)), {4, 5});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 500, 500, 0, true)
+                .apply(d));
+    REQUIRE(test_unitary_comparison(circ, d, true));
+    REQUIRE(d.count_n_qubit_gates(2) == 3);
+  }
+  GIVEN("Large circuit with ZZPhase") {
+    Circuit circ(6);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::X}, 0.3)), {0, 1});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::Y, Pauli::X}, 0.2)),
+        {0, 1, 2});
+    circ.add_box(
+        PauliExpCommutingSetBox({
+            {{Pauli::I, Pauli::Y, Pauli::I, Pauli::Z}, 1.2},
+            {{Pauli::X, Pauli::Y, Pauli::Z, Pauli::I}, 0.8},
+            {{Pauli::I, Pauli::I, Pauli::I, Pauli::Z}, 1.25},
+        }),
+        {1, 2, 3, 4});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Y, Pauli::X}, 0.1)), {2, 3});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::Y, Pauli::X}, 0.11)),
+        {1, 3, 4});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Y, Pauli::Y}, 0.2)), {4, 5});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::Z, Pauli::X}, 0.15)),
+        {2, 4, 5});
+    circ.add_box(
+        PauliExpBox(
+            SymPauliTensor({Pauli::X, Pauli::X, Pauli::X, Pauli::X}, 0.25)),
+        {2, 4, 5, 0});
+    circ.add_box(
+        PauliExpBox(
+            SymPauliTensor({Pauli::Y, Pauli::Z, Pauli::Z, Pauli::X}, 0.125)),
+        {1, 3, 5, 0});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 500, 500, 0, true)
+                .apply(d));
+    REQUIRE(test_unitary_comparison(circ, d, true));
+  }
+  GIVEN("Select TQE over ZZPhase") {
+    Circuit circ(3);
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::X, Pauli::Y}, 0.3)), {0, 1});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::Y, Pauli::Z}, 0.22)),
+        {0, 1, 2});
+    circ.add_box(
+        PauliExpBox(SymPauliTensor({Pauli::Z, Pauli::Y, Pauli::X}, 0.15)),
+        {0, 1, 2});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation(0.7, 0.3, 500, 500, 0, true)
+                .apply(d));
+    REQUIRE(test_unitary_comparison(circ, d, true));
+    // if the first XY was implemented using a ZZPhase
+    // then 2 TQEs is needed to conjugate the remaining two strings to weight 2
+    // hence 5 2-qubit gates in total.
+    REQUIRE(d.count_n_qubit_gates(2) == 4);
+  }
+}
+
+SCENARIO("Test GreedyPauliSimp for individual gates") {
+  Circuit circ(1);
+  circ.add_op<unsigned>(OpType::Z, {0});
+  std::vector<Op_ptr> ops_0q = {
+      get_op_ptr(OpType::Phase, 0.25),
+  };
+  std::vector<Op_ptr> ops_1q = {
+      get_op_ptr(OpType::noop),
+      get_op_ptr(OpType::Z),
+      get_op_ptr(OpType::X),
+      get_op_ptr(OpType::Y),
+      get_op_ptr(OpType::S),
+      get_op_ptr(OpType::V),
+      get_op_ptr(OpType::Sdg),
+      get_op_ptr(OpType::Vdg),
+      get_op_ptr(OpType::H),
+      get_op_ptr(OpType::Rz, 0.25),
+      get_op_ptr(OpType::Rz, 0.5),
+      get_op_ptr(OpType::Rx, 1),
+      get_op_ptr(OpType::Rx, 0.15),
+      get_op_ptr(OpType::Ry, 0.25),
+      get_op_ptr(OpType::Ry, -0.5),
+      get_op_ptr(OpType::PhasedX, {0.15, 0.2}),
+      get_op_ptr(OpType::PhasedX, {0.5, -0.5}),
+      get_op_ptr(OpType::PhasedX, {0.2, 1}),
+      get_op_ptr(OpType::T),
+      get_op_ptr(OpType::Tdg),
+  };
+  std::vector<Op_ptr> ops_2q = {
+      get_op_ptr(OpType::SWAP),
+      get_op_ptr(OpType::CX),
+      get_op_ptr(OpType::CY),
+      get_op_ptr(OpType::CZ),
+      get_op_ptr(OpType::ZZMax),
+      get_op_ptr(OpType::ZZPhase, 0.25),
+      get_op_ptr(OpType::ZZPhase, 0.5),
+      get_op_ptr(OpType::PhaseGadget, 0.5, 2),
+      get_op_ptr(OpType::XXPhase, 0.25),
+      get_op_ptr(OpType::XXPhase, 0.5),
+      get_op_ptr(OpType::YYPhase, 0.25),
+      get_op_ptr(OpType::YYPhase, 1),
+  };
+  for (Op_ptr op : ops_0q) {
+    Circuit circ(1);
+    circ.add_op<unsigned>(op, {});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(test_unitary_comparison(circ, d, true));
+  }
+  for (Op_ptr op : ops_1q) {
+    Circuit circ(1);
+    circ.add_op<unsigned>(op, {0});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(test_unitary_comparison(circ, d, true));
+  }
+  for (Op_ptr op : ops_2q) {
+    Circuit circ(2);
+    circ.add_op<unsigned>(op, {0, 1});
+    Circuit d(circ);
+    REQUIRE(Transforms::greedy_pauli_optimisation().apply(circ));
+    REQUIRE(test_unitary_comparison(circ, d, true));
   }
 }
 SCENARIO("Test GreedyPauliSimp pass construction") {